+++ nodes.rb
@@ -1,4 +1,4 @@
-module DOMReader
+module DOMParser
   
   module Nodes
